+++ CHANGES.md
@@ -5,6 +5,7 @@
 * Add check for package updates to menu item action "Check for Updates"
 * Add known failures are cleared for a fresh check when "Check for Updates" is used
 * Add FileSharingTalk nzb provider
+* Change optimize .png images to improve file/transfer size
 * Add tz version info to the about page
 * Change auto-install Cheetah dependency on first time installations (tested on Win)
 * Change add cryptography to recommended.txt